Add files via upload
Easiest way to edit this is to do each antigen from top to bottom by entering 0 when necessary and pressing enter to skip the row. This should take no more than 5 minutes if done at a decent pace. 

If you don't have a specific antigen in your panel that is listed, I recommend just deleting the entire column to prevent any unwanted errors.